# packages/extension/demos/openapi/todo-api.openapi.yaml
openapi: 3.0.3
info:
    title: Todo API (demo)
    description: Local demo API secured with x-api-key (stateless HTTP MCP pass-through).
    version: 0.1.0
servers:
    - url: http://127.0.0.1:3852
paths:
    /categories:
        get:
            operationId: list-categories
            summary: List todo categories
            security:
                - apiKeyAuth: []
            responses:
                '200':
                    description: Category list
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    categories:
                                        type: array
                                        items:
                                            type: object
                                            properties:
                                                id:
                                                    type: string
                                                name:
                                                    type: string
                                                color:
                                                    type: string
                '401':
                    description: Missing or invalid API key
    /categories/{categoryId}/todos:
        get:
            operationId: list-todos-by-category
            summary: List todos in a category
            security:
                - apiKeyAuth: []
            parameters:
                - name: categoryId
                  in: path
                  required: true
                  schema:
                      type: string
                - name: status
                  in: query
                  required: false
                  schema:
                      type: string
                      enum: [open, done]
            responses:
                '200':
                    description: Todos for category
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    categoryId:
                                        type: string
                                    todos:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/Todo'
                '401':
                    description: Missing or invalid API key
                '404':
                    description: Unknown category
    /todos:
        get:
            operationId: list-todos
            summary: List todos
            security:
                - apiKeyAuth: []
            parameters:
                - name: status
                  in: query
                  required: false
                  schema:
                      type: string
                      enum: [open, done]
                - name: categoryId
                  in: query
                  required: false
                  schema:
                      type: string
            responses:
                '200':
                    description: Todo list
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    todos:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/Todo'
                '401':
                    description: Missing or invalid API key
    /todos/{todoId}:
        get:
            operationId: get-todo
            summary: Get one todo by id
            security:
                - apiKeyAuth: []
            parameters:
                - name: todoId
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                '200':
                    description: Single todo
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    todo:
                                        $ref: '#/components/schemas/Todo'
                '401':
                    description: Missing or invalid API key
                '404':
                    description: Todo not found
components:
    securitySchemes:
        apiKeyAuth:
            type: apiKey
            in: header
            name: x-api-key
    schemas:
        Todo:
            type: object
            properties:
                id:
                    type: string
                title:
                    type: string
                status:
                    type: string
                categoryId:
                    type: string
                dueDate:
                    type: string